response to PR comment
